calculator: move controller to own directory
