+++ calculator/src/main/java/steammachinist/controller/CalculatorApi.java
@@ -1,4 +1,4 @@
-package steammachinist;
+package steammachinist.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
