+++ calculator/src/main/java/steammachinist/controller/CalculatorController.java
@@ -1,4 +1,4 @@
-package steammachinist;
+package steammachinist.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
